refactor: Update namespaces and reorganize CSV parser interfaces for better structure
