+++ src/SUI.Client/SUI.Client.StorageProcessFunction/StorageProcessFunctionOptions.cs
@@ -1,5 +1,3 @@
-using Shared;
-
 namespace SUI.StorageProcessFunction;
 
 public sealed class StorageProcessFunctionOptions
@@ -10,9 +8,6 @@ public sealed class StorageProcessFunctionOptions
 
     public string ProcessedContainerName { get; init; } = "processed";
     public string? MatchApiBaseAddress { get; init; }
-    public string SearchStrategy { get; init; } =
-        SharedConstants.SearchStrategy.Strategies.Strategy4;
-    public int? StrategyVersion { get; init; } = 2;
     public required string CsvParserName { get; init; }
 
     public static class CsvParserNameConstants
