+++ src/SUI.Client/SUI.Client.StorageProcessFunction/local.settings.sample.json
@@ -4,10 +4,10 @@
     "AzureWebJobsStorage": "UseDevelopmentStorage=true",
     "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
     "QueueName": "storage-process-job",
+    "PersonMatching:SearchStrategy": "strategy4",
+    "PersonMatching:StrategyVersion": "2",
     "StorageProcessFunction:ProcessedContainerName": "processed",
     "StorageProcessFunction:MatchApiBaseAddress": "http://localhost:5000",
-    "StorageProcessFunction:SearchStrategy": "strategy4",
-    "StorageProcessFunction:StrategyVersion": "2",
     "StorageProcessFunction:CsvParserName": "TypeOne"
   }
 }
