+++ tests/Unit.Tests/StorageProcessFunction/BlobFileOrchestratorTests.cs
@@ -3,6 +3,8 @@ using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Time.Testing;
 using Moq;
 using Shared.Models;
+using SUI.Client.Core.Application.Interfaces;
+using SUI.Client.Core.Infrastructure.CsvParsers;
 using SUI.StorageProcessFunction;
 using SUI.StorageProcessFunction.Application;
 using SUI.StorageProcessFunction.Application.Interfaces;
