+++ tests/Unit.Tests/StorageProcessFunction/PersonSpecificationCsvParserTests.cs
@@ -1,4 +1,4 @@
-using SUI.StorageProcessFunction.Infrastructure.Csv;
+using SUI.Client.Core.Infrastructure.CsvParsers;
 
 namespace Unit.Tests.StorageProcessFunction;
